update mts commands
tcpborphserver3 is updated to support calls to also target the dac. These
changes reflect the new api commands requiring the converter type for the remote
server to determine which mts configuration object to configure and run.

Updates the commands to also reflect changes to the mts marker debug information
and being able to get and set a target latency for mts runs.